Move pydantic validation models into separate file
